Move in memory index outside a specific test
